allow creating containers/vols w/ fixed handles
I'm a bit hesitant to do this, because I intend to only use these new
methods in tests (at least for now), but this commit adds a method for
creating volumes with a fixed handle rather than a random UUID, and a
new ContainerOwner for referencing/creating Containers with a fixed
handle.

the purpose of this is so that we can develop a fluent API for writing
tests for the new worker implementation that reference
containers/volumes with deterministic and semantic names.

while it makes the tests slower, I wanted to use the actual DB for these
worker tests so it's as close to real conditions as possible. I
considered writing hand-built fakes, but with the way the DB interfaces
are structured now, that would be a huge pain

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>