backport gardenruntime worker tests
these tests make use of the new gardenruntimetest and workertest
packages for higher level tests that should hopefully give us more
confidence than the existing atc/worker tests, and should be much more
resilient to refactors since the only interfaces we interact with are
the things we actually depend on, rather than interfaces we created just
for ease of testing. these tests are also just much more pleasant to
write IMO - it's cool to set up a scenario where you have some set of
workers, each with some set of containers and volumes, and exercise the
full system

the downside to these tests is that it's a lot harder to get high test
coverage for the error cases, but personally, I feel that the tradeoff
is worth it

also, worth noting that I wanted to use the stdlib testing package for
these tests in favour of Ginkgo, since I find the deeply nested
Context's that Ginkgo encourages result in tests that are ultimately
harder to read/write/modify, but I was "forced" to use Ginkgo since I
wanted to use the existing postgresrunner package which only currently
works with Ginkgo. so, I decided to write the tests as if I didn't have
BeforeEach/Context available to me, and personally I really like how
they turned out - the tests are all self-contained, so you don't need to
jump around the file to see why a test is failing.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>